+++ R/op-if.R
@@ -7,6 +7,19 @@ OpIf <- new_Op("OpIf", "if")
 #' @export
 infer_types_if <- function(pred, true_branch, false_branch) {
   assert_vt_has_ttype(pred, "bool", shape = integer())
+
+  # (C1)
+  for (nm in c("true_branch", "false_branch")) {
+    branch <- if (nm == "true_branch") true_branch else false_branch
+    n <- length(branch$inputs)
+    if (n != 0L) {
+      cli_abort(c(
+        "{.arg {nm}} must not have inputs.",
+        x = "Got {n} input{?s}."
+      ))
+    }
+  }
+
   out_types1 <- ValueTypes(func_output_types(true_branch))
   out_types2 <- ValueTypes(func_output_types(false_branch))
   if (length(out_types1) != length(out_types2)) {