+++ R/op-shift_right_arithmetic.R
@@ -9,7 +9,7 @@ OpShiftRightArithmetic <- new_Op(
 #' @rdname hlo_shift_right_arithmetic
 #' @export
 # fmt: skip
-infer_types_shift_right_arithmetic <- infer_types_integerish_biv # nolint
+infer_types_shift_right_arithmetic <- infer_types_integer_biv # nolint
 
 # fmt: skip
 hlo_shift_right_arithmetic_impl <- hlo_fn( # nolint